Moved ops to 'dev' module
