--- lib/on_container/bundler_ops.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module OnContainer
-  module BundlerOps
-    def ensure_project_gems_are_installed
-      system 'bundle check || bundle install'
-    end
-  end
-end
