+++ lib/on_container/dev/node_modules_ops.rb
@@ -1,0 +1,42 @@
+# frozen_string_literal: true
+
+require 'etc'
+
+module OnContainer
+  module Dev
+    module NodeModulesOps
+      APP_PATH = File.expand_path '.'
+  
+      def node_modules_path
+        "#{APP_PATH}/node_modules"
+      end
+  
+      def node_modules_owner_id
+        File.stat(node_modules_path).uid
+      end
+  
+      def current_user_id
+        Etc.getpwuid.uid
+      end
+  
+      def node_modules_belong_to_current_user?
+        node_modules_owner_id == current_user_id
+      end
+      
+      def make_node_modules_belong_to_current_user
+        target_ownership = "#{current_user_id}:#{current_user_id}"
+        system "sudo chown -R #{target_ownership} #{node_modules_path}"
+      end
+      
+      def ensure_node_modules_belong_to_current_user
+        return if node_modules_belong_to_current_user?
+      
+        make_node_modules_belong_to_current_user
+      end
+  
+      def ensure_project_node_packages_are_installed
+        system 'yarn check --integrity || yarn install'
+      end
+    end
+  end
+end
